Decompose behavior_tests into 7 focused files
Extract query_tests, policy_tests, lifecycle_tests, negative_tests,
endpoint_tests, domain_lookups into smoke_test/ subdirectory.
Also moves event_checks and form_submission to correct paths.

Phase 3 of HEC-493.
